ZNTA-2056 split tests into separate classes
